Move executor factory from ats/storage to ats/setup
storage had no business importing ax and alias — it's a persistence
layer that was also assembling the query engine on top of itself.
The factory is pure wiring (70 lines, zero logic), so moving it
breaks no behavior. Callers change their import path; storage drops
2 import edges.
